requirements.txt
python-telegram-bot==12.7
airtable-python-wrapper=0.12.0